Fix for adding a todo item
When adding a new item it would not show in the list. This was because
there was two TodoHelper instances. One that belonged to the
TodosRecyclerViewActivity and another one created in
AddOrUpdateTodoActivity. So when we added a new todo it was adding it to
the latter instead of the former.

However, we could see the new todo when relaunching the app because in
both cases the todo was saved so it was restored by
TodosRecyclerViewActivity. The solution is to make sure there is only
one TodoHelper which manages the only list of todos. Updating the todo
list isn't magic, we still need to use notifyDataSetChanged() for the
UI.

This also fixes a bug where the first ever load (when there are no
todos) would cause the app to crash because theres no JSON to parse.